pyproject.toml
[project]
name = "purchase-agency-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.14",
    "beautifulsoup4>=4.13.3",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "lxml>=5.3.1",
    "openai>=1.68.2",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlmodel>=0.0.16",
    "uvicorn>=0.34.0",
]
